# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  model.RoleVO:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      sort_order:
        type: integer
      status:
        type: integer
      status_name:
        type: string
    type: object
...
  response.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.CreateRoleRequest:
    properties:
      code:
        maxLength: 32
        minLength: 2
        type: string
      description:
        maxLength: 256
        type: string
      name:
        maxLength: 64
        minLength: 2
        type: string
      sort_order:
        type: integer
      status:
        type: integer
    required:
    - code
    - name
    type: object
...
  service.UpdateRoleRequest:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      sort_order:
        type: integer
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@huaan-medical.com
    name: API Support
  description: 华安医疗门诊预约系统后端API接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 华安医疗预约系统 API
  version: "1.0"
paths:
...
  /api/admin/roles:
    get:
      consumes:
      - application/json
      description: 分页查询角色列表
      parameters:
      - description: 页码
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        required: true
        type: integer
      - description: 关键词(编码/名称)
        in: query
        name: keyword
        type: string
      - description: 状态 0禁用 1启用
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/model.RoleVO'
                        type: array
                    type: object
              type: object
      security:
      - BearerAdmin: []
      summary: 角色列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 创建角色
      parameters:
      - description: 角色信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.RoleVO'
              type: object
      security:
      - BearerAdmin: []
      summary: 创建角色
      tags:
      - 角色管理
  /api/admin/roles/{id}:
    put:
      consumes:
      - application/json
      description: 更新角色信息
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.RoleVO'
              type: object
      security:
      - BearerAdmin: []
      summary: 更新角色
      tags:
      - 角色管理
...
securityDefinitions:
  Bearer:
    description: 请在Token前加上"Bearer "前缀
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"